+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - 🔢 Multi-model viewer to see multiple Gaussian models, enabled by `multi-model` feature.
 
+### Changed
+
+- ⏫ Increase the capability of the viewer to handle larger models.
+
 ## [0.1.0](https://crates.io/crates/wgpu-3dgs-viewer/0.1.0) - 2025-02-27
 
 The first version of this project.